Refactor cli module into more files
